[v0.89.1][maintainability] Split provider/http_family.rs by responsibility (Closes #1960)
